work in progress...
got data including on page relaod...

made a __layout inside of projects and imported PrevNextProject
component inside of there as opposed to inside every project.md.

have all the data and fixed the not navigating issue with
e.stopPropagation.

working on this components nav logic for the prev and next buttons for
whan there are more than 3... want it to be further scalable