adding some polish to the ...
success and error modals on form submission on the frontend...

moved the Loader more into the form.

I like the idea of a universal loader but at present only time loader
will show up is on form submission and it makes more since for it to be
inside the form at least for the moment.

adjested some media queries and the modal text.